feat: Prevent deletion of the segment if it still has sensors

+++ app/database/segments_crud.py
@@ -79,6 +79,12 @@ def delete_segment_by_id(session: Session, segment_id: int):
             status_code=status.HTTP_404_NOT_FOUND
         )
     
+    if segment.sensors:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail='Segment cannot be deleted while it still has sensors'
+        )
+    
     session.delete(segment)
     session.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
